move sasl utility to ircutils
